+++ templates/part.audio.php
@@ -71,6 +71,11 @@
                    step="0.02" value="1">
         </div>
 
+        <div class="sm2-inline-element sm2-button-element" data-placement="left"
+             title="<?php p($l->t('Playback speed')); ?>">
+            <div id="playerSpeed" class="sm2-button-bd sm2-inline-button speed">1x</div>
+        </div>
+
         <div class="sm2-inline-element sm2-button-element sm2-repeat" style="left: 7px;" data-placement="left"
              title="<?php p($l->t('Repeat title/list')); ?>">
             <div id="playerRepeat" class="sm2-button-bd sm2-inline-button repeat">